+++ src/main/java/Rules/QuotientRule.java
@@ -47,7 +47,7 @@ public class QuotientRule extends DerivationRule {
     }
 
     @Override
-    public int getResult(List<Integer> dims) {
+    public double getResult(List<Integer> dims) {
         return 0;
     }
 }
